Prepare v0.4 ALPHA
The final alpha 0.4 version will address the following missing stuff:
- MMC
- WDI
- DiskNVCache
- Backup
- DWM
- PCA
- AppHVSI
- FDE
- WD
- Search
- TPM
- MMCSnapIns
- Taskbar
- Netlogon
- Firewall (completed)
- ICM
- GPO preferences
- UserExperienceVirtualization
- inetres
- Tablet related stuff 

Once this is done I can roll out ALPHA 0.4 (final version) and work on the Beta which is focused only on bug fixing (see "fixme"). 

This then is the most complete script ever written when it comes to Windows, I checked all GPO's (18.726 incl hidden ones).

Once the bug fixing then is complete I will roll out the final version, which is scheduled for Dec./Jan. When the final version is done I can go ahead and address possible user related findings and add workarounds or so, which is already a mentioned in the to-do for v1.1.